refactor(headless): extract permission-intervention decision engine
Move handlePermissionIntervention and its policy/types/helpers into
permission-intervention.ts beside permission-grants.ts so task-agent-controller
keeps only the main and repair call sites.

Closes #1410
